requirements.txt
<<<<<<< HEAD
asgiref==3.9.1
Django==5.2.5
python-dotenv==1.1.1
sqlparse==0.5.3
=======
# ==== Núcleo ====
Django~=5.1

# ==== Deploy (Render/Heroku/Ubuntu) ====
gunicorn~=21.2           # servidor WSGI para produção
whitenoise[brotli]~=6.6  # servir arquivos estáticos (CSS/JS/Imagens) direto do app

# ==== Banco de dados e config por variável de ambiente ====
dj-database-url~=2.1     # parse de DATABASE_URL (opcional mas prático)
python-dotenv~=1.0       # ler .env localmente (útil no dev)

# Se for usar PostgreSQL (Render costuma usar):
psycopg2-binary~=2.9

# ==== Imagens e uploads ====
Pillow~=10.4             # manipulação básica de imagens (thumbnails, formatos, etc.)

# ==== Bootstrap (via Django helper, opcional) ====
# Você pode usar Bootstrap como arquivos estáticos puros.
# Se preferir helpers de template/templatetags:
django-bootstrap5~=24.3

# ==== Otimização/Caching de estáticos (opcional) ====
# Úteis para minificar CSS/JS caso use django-compressor
django-compressor~=4.5
rcssmin~=1.1.2
rjsmin~=1.2.1

# ==== Segurança/Qualidade (opcionais, mas recomendados) ====
# Sentry para monitorar erros em produção:
# sentry-sdk~=2.7
# Health check endpoint simples:
# django-health-check~=3.18.3

# ==== Internacionalização em Windows (se precisar) ====
# tzdata~=2024.1

# ==== Observações ====
# 1) Para Bootstrap/Carrossel: coloque os arquivos estáticos (bootstrap.min.css/js) em /static
#    ou use CDN no template. O pacote django-bootstrap5 é opcional.
# 2) Mapa interativo:
#    - Via <iframe> do Google Maps: não precisa de pacote Python.
#    - Via LeafletJS (recomendado para interativo): apenas inclua os arquivos JS/CSS do Leaflet em /static ou CDN.
#      (django-leaflet é opcional e mais útil quando há formulários/map widgets; não é necessário só para exibir um mapa.)
>>>>>>> fd91b9f747a36440807c2d20af3f46aa92636eea
